% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendeemovementbreakdown.R
\name{attendeeMovementBreakdown}
\alias{attendeeMovementBreakdown}
\title{Prepare data to show people joining and leaving webinar}
\usage{
attendeeMovementBreakdown(timeData)
}
\arguments{
\item{timeData}{value returned from `getTimeData`}
}
\value{
A long version of the timeData
}
\description{
Prepare data to show people joining and leaving webinar
}
