% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getattendees.R
\name{getAttendees}
\alias{getAttendees}
\title{Pull list of attendees from default Zoom Attendee Report}
\usage{
getAttendees(file)
}
\arguments{
\item{file}{A string representing the filepath to the Zoom Attendee Report}
}
\value{
a tibble containing attendees of the webinar
}
\description{
Pull list of attendees from default Zoom Attendee Report
}
