% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcountryinfo.R
\name{getCountryInfo}
\alias{getCountryInfo}
\title{Summary statistics of countries whose citizens joined the webinar}
\usage{
getCountryInfo(attendeeData)
}
\arguments{
\item{attendeeData}{tibble output of getAttendees}
}
\value{
a tibble containing summary statistics of countries attending the event
}
\description{
Summary statistics of countries whose citizens joined the webinar
}
