% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataindex.R
\name{getDataIndex}
\alias{getDataIndex}
\title{Retrieve index of embedded data}
\usage{
getDataIndex(file, heading)
}
\arguments{
\item{file}{A string representing the filepath to the Zoom Attendee Report}

\item{heading}{A string used to identify the header of the data of interest}
}
\value{
an integer representing which line the data of interest begins
}
\description{
Retrieve index of embedded data
}
