% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmovementinfo.R
\name{getMovementInfo}
\alias{getMovementInfo}
\title{Isolate movements from attendee data}
\usage{
getMovementInfo(attendeeData, emailList = NA)
}
\arguments{
\item{attendeeData}{tibble output of getAttendees}

\item{emailList}{if not NA then a string vector of email address to keep in the analysis}
}
\value{
a tibble with times whih attendees left or joined the webinar.
}
\description{
Isolate movements from attendee data
}
