% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettimedata.R
\name{getTimeData}
\alias{getTimeData}
\title{Get information on time intervals in which attendees joined and left}
\usage{
getTimeData(intervals, attendeeData)
}
\arguments{
\item{intervals}{vector time interval output from `createIntervals`}

\item{attendeeData}{tibble output of getAttendees}
}
\value{
a tibble with summary information of time intervals throughout webinar
}
\description{
Get information on time intervals in which attendees joined and left
}
