% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getuniqueattendees.R
\name{getUniqueAttendees}
\alias{getUniqueAttendees}
\title{Combines dublicate data}
\usage{
getUniqueAttendees(attendeeData)
}
\arguments{
\item{attendeeData}{tibble output of getAttendees}
}
\value{
a tibble containing the total amount of time an attendee stayed in the conference.
}
\description{
Combines dublicate data
}
