% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showattendeemovement.R
\name{showAttendeeMovement}
\alias{showAttendeeMovement}
\title{Visualizations of timed data}
\usage{
showAttendeeMovement(timeData, type = "cumulative")
}
\arguments{
\item{timeData}{tibble output of getTimeData}

\item{type}{a string value of either "cumulative", "discrete", or "breakown"}
}
\value{
a ggplot
}
\description{
Visualizations of timed data
}
